refactor(macros): organize by purpose
